+++ lib/jsapi/model/attributes.rb
@@ -10,12 +10,18 @@ module Jsapi
       end
 
       def method_missing(*args) # :nodoc:
-        name = args.first
-        attribute?(name) ? self[name] : super
+        name = args.first.to_s
+        _attr_readers.key?(name) ? _attr_readers[name] : super
       end
 
       def respond_to_missing?(param1, _param2) # :nodoc:
-        attribute?(param1) ? true : super
+        _attr_readers.key?(param1.to_s) ? true : super
+      end
+
+      private
+
+      def _attr_readers
+        @_attr_readers ||= attributes.transform_keys(&:underscore)
       end
     end
   end